change qte to float

+++ report/forms.py
@@ -25,8 +25,8 @@ class ProductForm(BaseModelForm):
     designation = forms.CharField(widget=forms.TextInput(attrs=getAttrs('control', 'Désignation')))
     image = forms.ImageField(widget=forms.ClearableFileInput(attrs={'class': 'custom-file-input', 'accept': 'image/*'}), required=False)
     delais_expiration = forms.IntegerField(widget=forms.NumberInput(attrs=getAttrs('control', 'Délais d\'expiration')))
-    qte_per_cond = forms.IntegerField(widget=forms.NumberInput(attrs=getAttrs('control', 'Quantité par unité')))
-    qte_per_pal = forms.IntegerField(widget=forms.NumberInput(attrs=getAttrs('control', 'Quantité par palette')))
+    qte_per_cond = forms.FloatField(widget=forms.NumberInput(attrs=getAttrs('control', 'Quantité par unité')))
+    qte_per_pal = forms.FloatField(widget=forms.NumberInput(attrs=getAttrs('control', 'Quantité par palette')))
     family = forms.ModelChoiceField(queryset=Family.objects.all(), widget=forms.Select(attrs=getAttrs('control')), empty_label="Famille")
     packing = forms.ModelChoiceField(queryset=Packing.objects.all(), widget=forms.Select(attrs=getAttrs('control')), empty_label="Conditionnement")
     alert_stock = forms.IntegerField(widget=forms.NumberInput(attrs=getAttrs('control', 'Alerte Stock')))
@@ -37,8 +37,8 @@ class ProductForm(BaseModelForm):
         
 class MProductForm(BaseModelForm):
     
-    qte_per_cond = forms.IntegerField(widget=forms.NumberInput(attrs=getAttrs('control', 'Quantité par copnditionnement')))
-    qte_per_pal = forms.IntegerField(widget=forms.NumberInput(attrs=getAttrs('control', 'Quantité par palette')))
+    qte_per_cond = forms.FloatField(widget=forms.NumberInput(attrs=getAttrs('control', 'Quantité par copnditionnement')))
+    qte_per_pal = forms.FloatField(widget=forms.NumberInput(attrs=getAttrs('control', 'Quantité par palette')))
     packing = forms.ModelChoiceField(queryset=Packing.objects.all(), widget=forms.Select(attrs=getAttrs('control')), empty_label="Conditionnement")
     alert_stock = forms.IntegerField(widget=forms.NumberInput(attrs=getAttrs('control', 'Alerte Stock')))
